Allow targeting master node.
As master node may be exposed through a hostname and not an IP, using
the advertise address to find the join token doesn't always work.

This commit allows targetting the master node by specifying a standard
salt target. If a join token is found for any of the targetted nodes,
this join token will be used by the minion to join the cluster.

I guess this may also reduce the pressure on the master as it is
possible in practice to target all manager nodes and not only the
master.